refactor(types): one Scope type, u32 scope and pack indices
`Scope::Grp(u64)/Seq(u64)`, serialize's `CompWrap`, and the graph's
`Property` were three encodings of the same fact. `Scope { kind, index }`
with a shared `ScopeKind` is the one type; graphify no longer needs
projection helpers to pull the index and kind back out.

Scope and pack indices are `u32` like every other index in the pipeline.
`ObjNode::Pack` carrying a `u64` forced the object arena's node to 8-byte
alignment; with `u32` the node is 12 bytes instead of 24.

Byte-identical to the oracle (QCheck plus 9000 fuzz rounds).

+++ typeset/src/compiler/types/doc.rs
@@ -29,7 +29,7 @@ pub(crate) enum ObjNode {
     Grp(ObjId),
     Seq(ObjId),
     Nest(ObjId),
-    Pack(u64, ObjId),
+    Pack(u32, ObjId),
     Comp(ObjId, ObjId, bool),
 }
